// modules/meetings/ui/components/status-filter.tsx
//-- external imports
import {
  CircleXIcon,
  CircleCheckIcon,
  ClockArrowUpIcon,
  VideoIcon,
  LoaderIcon,
} from 'lucide-react'
//-- internal imports
import { MeetingStatus } from '../../types'
import { CommandSelect } from '@/components/command-select'
import { useMeetingsFilters } from '../../hooks/use-meetings-filters'

const options = [
  {
    id: MeetingStatus.Upcoming,
    value: MeetingStatus.Upcoming,
    children: (
      <div className='flex items-center gap-x-2 capitalize'>
        <ClockArrowUpIcon />
        {MeetingStatus.Upcoming}
      </div>
    ),
  },
  {
    id: MeetingStatus.Active,
    value: MeetingStatus.Active,
    children: (
      <div className='flex items-center gap-x-2 capitalize'>
        <VideoIcon />
        {MeetingStatus.Active}
      </div>
    ),
  },
  {
    id: MeetingStatus.Cancelled,
    value: MeetingStatus.Cancelled,
    children: (
      <div className='flex items-center gap-x-2 capitalize'>
        <CircleXIcon />
        {MeetingStatus.Cancelled}
      </div>
    ),
  },
  {
    id: MeetingStatus.Completed,
    value: MeetingStatus.Completed,
    children: (
      <div className='flex items-center gap-x-2 capitalize'>
        <CircleCheckIcon />
        {MeetingStatus.Completed}
      </div>
    ),
  },
  {
    id: MeetingStatus.Processing,
    value: MeetingStatus.Processing,
    children: (
      <div className='flex items-center gap-x-2 capitalize'>
        <LoaderIcon />
        {MeetingStatus.Processing}
      </div>
    ),
  },
]

export function StatusFilter() {
    const [filters, setFilters] = useMeetingsFilters()

    return (
        <CommandSelect
            options={options}
            placeholder='Status'
            className='h-9'
            value={filters.status ?? ''}
            onSelect={(value) => setFilters({status: value as MeetingStatus})}
        />
    )
}